..F.....S. [ZBX-24916] fixed date/time calculation for numeric (unsigned) item values with unixtime units after the year 2038
* commit 'd4e7b92b87aec5fbde9bfad98ab4351230a2d2d1':
  .D........ [ZBX-24916] fixed changelog
  .D........ [ZBX-24916] adjusted changelog
  ..F....... [ZBX-24916] adjusted full time display for large unix time
  ........S. [ZBX-24916] fixed wrong unixtime calculation after year 2038

+++ src/libs/zbxexpression/evalfunc.c
@@ -317,7 +317,7 @@ static void	add_value_suffix(char *value, size_t max_len, const char *units, uns
 		case ITEM_VALUE_TYPE_UINT64:
 			if (0 == strcmp(units, "unixtime"))
 			{
-				time = (time_t)atoi(value);
+				time = (time_t)atol(value);
 				local_time = localtime(&time);
 				strftime(value, max_len, "%Y.%m.%d %H:%M:%S", local_time);
 				break;
